correcoes no script

+++ Semestre 02/Banco de Dados/21.02.20Aula03/ExercicioAula03.sql
@@ -5,7 +5,7 @@ CREATE TABLE tb_departamento(
 );
 
 CREATE TABLE tb_projeto(
-    cd_projeto INTEGER,
+    cd_projeto NUMBER(5),
     nm_projeto VARCHAR2(300),
     dt_ini_projeto DATE,
     dt_fin_projeto DATE,
